requirements.txt
fastapi>=0.100.0
numpy>=1.23.0
pandas>=1.5.0
pydantic>=2.0.0
openpyxl>=3.1.0
playwright>=1.41.0
pyyaml>=6.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
networkx>=3.0
uvicorn>=0.23.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0